refaktoring from code

+++ ConsoleApp1/LoginApp/UserMethoden/UserInformation/UserProperties.cs
@@ -3,9 +3,9 @@
     public class UserProperties
     {
         public string Username { get; set; }
-        public int Password { get; set; }
+        public long Password { get; set; }
 
-        public UserProperties(string username, int password)
+        public UserProperties(string username, long password)
         {
             Username = username;
             Password = password;
